Add files via upload
Firebase Database
Task 1 (6 marks)
Add the following features to the World Cup application you developed in the previous week:

The app should export its database to other applications by implementing the content provider class.  (3m)
The content provider authority should be "NOV.app.YOUR_FIRST_NAME".
To test your work, develop another application, "App(B): Content Resolver", that should retrieve the current number of matches in the database and display it in a text view. (3m)
Note:  don't forget to change the authority in App(B)

Task 2 (4 marks)- Save Matches in the Cloud

The current implementation saves all the Matches in a local database. There is a need to push every new match to the cloud. 

When a new match score is added to the local database, the app must push that details to the Google firebase database. (2m)
If the Navigation menu option "Remove All Matches" is selected, all matches in the Google firebase database must be removed. (2m)